# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Resumo API",
    "version": "1.0.0",
    "description": "API para geração de resumos, perguntas e avaliação de respostas"
  },
  "paths": {
    "/resumos": {
      "post": {
        "summary": "Gerar resumo e perguntas com base em um tema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tema": {
                    "type": "string"
                  }
                },
                "required": ["tema"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resumo e perguntas gerados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resumo"
                }
              }
            }
          },
          "400": {
            "description": "Tema não informado"
          },
          "500": {
            "description": "Erro ao gerar conteúdo"
          }
        }
      },
      "get": {
        "summary": "Listar todos os resumos",
        "responses": {
          "200": {
            "description": "Lista de resumos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resumo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/resumos/{id}": {
      "put": {
        "summary": "Editar resumo existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Resumo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resumo editado com sucesso"
          },
          "404": {
            "description": "Resumo não encontrado"
          },
          "500": {
            "description": "Erro ao editar resumo"
          }
        }
      },
      "delete": {
        "summary": "Deletar resumo existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumo deletado com sucesso"
          },
          "404": {
            "description": "Resumo não encontrado"
          },
          "500": {
            "description": "Erro ao deletar resumo"
          }
        }
      }
    },
    "/teste": {
      "post": {
        "summary": "Enviar respostas e calcular pontuação",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Resultado"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado salvo e pontuação retornada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resultado"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "500": {
            "description": "Erro interno ao salvar resultado"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Pergunta": {
        "type": "object",
        "properties": {
          "numero": { "type": "integer" },
          "pergunta": { "type": "string" },
          "opcoes": {
            "type": "array",
            "items": { "type": "string" }
          },
          "correta": { "type": "string" }
        }
      },
      "Resumo": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "tema": { "type": "string" },
          "resumo": { "type": "string" },
          "perguntas": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Pergunta" }
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Resultado": {
        "type": "object",
        "properties": {
          "aluno": { "type": "string" },
          "tema": { "type": "string" },
          "resumo": { "type": "string" },
          "perguntas": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Pergunta" }
          },
          "respostas": {
            "type": "array",
            "items": { "type": "string" }
          },
          "pontuacao": { "type": "number" },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["aluno", "tema", "resumo", "perguntas", "respostas"]
      }
    }
  }
}